# docs/contracts/social-media-v2-openapi.json
{
  "components": {
    "schemas": {
...
      "AuthMeResponse": {
        "properties": {
          "access_mode": {
            "title": "Access Mode",
            "type": "string"
          },
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          },
          "brand_id": {
            "title": "Brand Id",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "is_internal_staff": {
            "title": "Is Internal Staff",
            "type": "boolean"
          },
          "revoked": {
            "title": "Revoked",
            "type": "boolean"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "settings_visible": {
            "title": "Settings Visible",
            "type": "boolean"
          },
          "source_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source System"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "authenticated",
          "user_id",
          "email",
          "source_system",
          "brand_id",
          "role",
          "access_mode",
          "settings_visible",
          "is_internal_staff",
          "expires_at",
          "revoked"
        ],
        "title": "AuthMeResponse",
        "type": "object"
      },
...
      "BrandScope": {
        "properties": {
          "requested_brand_id": {
            "title": "Requested Brand Id",
            "type": "string"
          },
          "resolved_brand_ids": {
            "items": {
              "type": "string"
            },
            "title": "Resolved Brand Ids",
            "type": "array"
          },
          "rollup": {
            "title": "Rollup",
            "type": "boolean"
          }
        },
        "required": [
          "requested_brand_id",
          "rollup",
          "resolved_brand_ids"
        ],
        "title": "BrandScope",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationsReadinessResponse": {
        "properties": {
          "database_configured": {
            "title": "Database Configured",
            "type": "boolean"
          },
          "platforms": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ReadinessPlatform"
            },
            "title": "Platforms",
            "type": "array"
          },
          "runtime_mode": {
            "$ref": "#/components/schemas/RuntimeMode"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "writes_enabled": {
            "title": "Writes Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "status",
          "runtime_mode",
          "writes_enabled",
          "database_configured"
        ],
        "title": "OperationsReadinessResponse",
        "type": "object"
      },
...
      "PlatformId": {
        "enum": [
          "facebook",
          "instagram",
          "tiktok"
        ],
        "title": "PlatformId",
        "type": "string"
      },
      "ReadinessPlatform": {
        "properties": {
          "account_count": {
            "title": "Account Count",
            "type": "integer"
          },
          "last_sync_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Sync At"
          },
          "pending_job_count": {
            "title": "Pending Job Count",
            "type": "integer"
          },
          "platform": {
            "$ref": "#/components/schemas/PlatformId"
          }
        },
        "required": [
          "platform",
          "account_count",
          "last_sync_at",
          "pending_job_count"
        ],
        "title": "ReadinessPlatform",
        "type": "object"
      },
...
      "RuntimeMode": {
        "enum": [
          "development",
          "dormant",
          "cutover_read_only",
          "cutover_credential_migration",
          "cutover_canary",
          "cutover_control_plane_drain",
          "cutover_activation",
          "active"
        ],
        "title": "RuntimeMode",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Social Media V2",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/logout": {
      "post": {
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "social_media_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Social Media Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout"
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "auth_me_api_auth_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "social_media_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Social Media Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Auth Me"
      }
    },
...
    "/api/media/{platform}/{content_id}": {
      "get": {
        "operationId": "social_media_api_media__platform___content_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PlatformId"
            }
          },
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "title": "Content Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "brand_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand Id"
            }
          },
          {
            "in": "query",
            "name": "rollup",
            "required": false,
            "schema": {
              "default": false,
              "title": "Rollup",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "cookie",
            "name": "social_media_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Social Media Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Social Media"
      }
    },
...
    "/api/operations/readiness": {
      "get": {
        "operationId": "readiness_api_operations_readiness_get",
        "parameters": [
          {
            "in": "query",
            "name": "brand_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brand Id"
            }
          },
          {
            "in": "query",
            "name": "rollup",
            "required": false,
            "schema": {
              "default": false,
              "title": "Rollup",
              "type": "boolean"
            }
          },
          {
            "in": "cookie",
            "name": "social_media_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Social Media Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dormant readiness probe",
        "tags": [
          "operations"
        ]
      }
    },
...
  }
}